+++ api_auth0_fga.go
@@ -33,33 +33,32 @@ type Auth0FgaApi interface {
 
 	/*
 		 * Check Check whether a user is authorized to access an object
-		 * The check API will return whether the user has a certain relationship with an object in a certain store.
-	Path parameter `store_id` as well as the body parameter `tuple_key` with specified `object`, `relation` and `user` subfields are all required.
-	Optionally, a `contextual_tuples` object may also be included in the body of the request. This object contains one field `tuple_keys`, which is an array of tuple keys.
+		 * The Check API queries to check if the user has a certain relationship with an object in a certain store.
+	A `contextual_tuples` object may also be included in the body of the request. This object contains one field `tuple_keys`, which is an array of tuple keys.
+	You may also provide an `authorization_model_id` in the body. This will be used to assert that the input `tuple_key` is valid for the model specified. If not specified, the assertion will be made against the latest authorization model ID.
 	The response will return whether the relationship exists in the field `allowed`.
 
-	## [Limits](https://docs.fga.dev/intro/dashboard#limitations)
 	## Example
-	In order to check if user `anne@auth0.com` has a `can_read` relationship with object `document:2021-budget` given the following contextual tuple
+	In order to check if user `user:anne` of type `user` has a `reader` relationship with object `document:2021-budget` given the following contextual tuple
 	```json
 	{
-	  "user": "anne@auth0.com",
+	  "user": "user:anne",
 	  "relation": "member",
 	  "object": "time_slot:office_hours"
 	}
 	```
-	a check API call should be fired with the following body:
+	the Check API can be used with the following request body:
 	```json
 	{
 	  "tuple_key": {
-	    "user": "anne@auth0.com",
-	    "relation": "can_read",
+	    "user": "user:anne",
+	    "relation": "reader",
 	    "object": "document:2021-budget"
 	  },
 	  "contextual_tuples": {
 	    "tuple_keys": [
 	      {
-	        "user": "anne@auth0.com",
+	        "user": "user:anne",
 	        "relation": "member",
 	        "object": "time_slot:office_hours"
 	      }
@@ -81,21 +80,13 @@ type Auth0FgaApi interface {
 
 	/*
 		 * Expand Expand all relationships in userset tree format, and following userset rewrite rules.  Useful to reason about and debug a certain relationship
-		 * The expand API will return all users (including user and userset) that have certain relationship with an object in a certain store.
-	This is different from the `/stores/{store_id}/read` API in that both users and computed references are returned.
-	Path parameter `store_id` as well as body parameter `object`, `relation` are all required.
-	The response will return a userset tree whose leaves are the user id and usersets.  Union, intersection and difference operator are located in the intermediate nodes.
+		 * The Expand API will return all users and usersets that have certain relationship with an object in a certain store.
+	This is different from the `/stores/{store_id}/read` API in that both users and computed usersets are returned.
+	Body parameters `tuple_key.object` and `tuple_key.relation` are all required.
+	The response will return a tree whose leaves are the specific users and usersets. Union, intersection and difference operator are located in the intermediate nodes.
 
-	## [Limits](https://docs.fga.dev/intro/dashboard#limitations)
 	## Example
-	Assume the following type definition for document:
-	```yaml
-	  type document
-	    relations
-	      define reader as self or writer
-	      define writer as self
-	```
-	In order to expand all users that have `reader` relationship with object `document:2021-budget`, an expand API call should be fired with the following body
+	To expand all users that have the `reader` relationship with object `document:2021-budget`, use the Expand API with the following request body
 	```json
 	{
 	  "tuple_key": {
@@ -104,7 +95,7 @@ type Auth0FgaApi interface {
 	  }
 	}
 	```
-	Auth0 FGA's response will be a userset tree of the users and computed usersets that have read access to the document.
+	Auth0 FGA's response will be a userset tree of the users and usersets that have read access to the document.
 	```json
 	{
 	  "tree":{
@@ -117,7 +108,7 @@ type Auth0FgaApi interface {
 	            "leaf":{
 	              "users":{
 	                "users":[
-	                  "bob@auth0.com"
+	                  "user:bob"
 	                ]
 	              }
 	            }
@@ -149,11 +140,34 @@ type Auth0FgaApi interface {
 	ExpandExecute(r ApiExpandRequest) (ExpandResponse, *_nethttp.Response, error)
 
 	/*
-	 * ListObjects [EXPERIMENTAL] Returns a list of all of the object IDs of the provided type that the given user has a specific relation with
-	 * Please see https://openfga.dev/docs.
-	 * @param ctx _context.Context - for authentication, logging, cancellation, deadlines, tracing, etc. Passed from http.Request or context.Background().
-	 * @return ApiListObjectsRequest
-	 */
+		 * ListObjects [EXPERIMENTAL] Get all objects of the given type that the user has a relation with
+		 * The ListObjects API returns a list of all the objects of the given type that the user has a relation with. To achieve this, both the store tuples and the authorization model are used.
+	An `authorization_model_id` may be specified in the body. If it is, it will be used to decide the underlying implementation used. If it is not specified, the latest authorization model ID will be used.
+	You may also specify `contextual_tuples` that will be treated as regular tuples.
+	The response will contain the related objects in an array in the "objects" field of the response and they will be strings in the object format `<type>:<id>` (e.g. "document:roadmap")
+	## Example
+	In order to list the objects of type document that user `user:anne` has a `reader` relationship with, while passing the Anne is an editor of the marketing folder in the contextual tuples, You can issue a ListObjects API request that includes the contextual tuples:
+	```json
+	{
+	  "authorization_model_id": "01G5JAVJ41T49E9TT3SKVS7X1J",
+	  "user": "user:anne",
+	  "relation": "reader",
+	  "type": "document"
+	  "contextual_tuples": {
+	    "tuple_keys": [
+	      {
+	        "user": "user:anne",
+	        "relation": "editor",
+	        "object": "folder:marketing"
+	      }
+	    ]
+	  }
+	}
+	```
+	Auth0 FGA's response will be of the format: `{ "objects": ["document:roadmap"] }` and include document Anne is related to as reader
+		 * @param ctx _context.Context - for authentication, logging, cancellation, deadlines, tracing, etc. Passed from http.Request or context.Background().
+		 * @return ApiListObjectsRequest
+	*/
 	ListObjects(ctx _context.Context) ApiListObjectsRequest
 
 	/*
@@ -164,18 +178,18 @@ type Auth0FgaApi interface {
 
 	/*
 		 * Read Get tuples from the store that matches a query, without following userset rewrite rules
-		 * The POST read API will return the tuples for a certain store that matches a query filter specified in the body. Tuples and type definitions allow Auth0 FGA to determine whether a relationship exists between an object and an user.
-	It is different from the `/stores/{store_id}/expand` API in that only read returns relationship tuples that are stored in the system and satisfy the query.
-	It does not expand or traverse the graph by taking the authorization model into account.Path parameter `store_id` is required.  In the body:
-	1. Object is mandatory. An object can be a full object (e.g., `type:object_id`) or type only (e.g., `type:`).
-	2. User is mandatory in the case the object is type only.
+		 * The Read API will return the tuples for a certain store that match a query filter specified in the body of the request. It is different from the `/stores/{store_id}/expand` API in that it only returns relationship tuples that are stored in the system and satisfy the query.
+	In the body:
+	1. `tuple_key.object` is mandatory. It can be a full object (e.g., `type:object_id`) or type only (e.g., `type:`).
+	2. `tuple_key.user` is mandatory in the case the `tuple_key.object` is a type only.
+	3. `authorization_model_id` is optional. If specified, it will be used to assert that the input `tuple_key` is valid for the model specified. If not specified, the latest authorization model ID will be used.
 	## Examples
 	### Query for all objects in a type definition
-	To query for all objects that `bob@auth0.com` has `reader` relationship in the document type definition, call read API with body of
+	To query for all objects that `user:bob` has `reader` relationship in the document type definition, call read API with body of
 	```json
 	{
 	 "tuple_key": {
-	     "user": "bob@auth0.com",
+	     "user": "user:bob",
 	     "relation": "reader",
 	     "object": "document:"
 	  }
@@ -187,7 +201,7 @@ type Auth0FgaApi interface {
 	  "tuples": [
 	    {
 	      "key": {
-	        "user": "bob@auth0.com",
+	        "user": "user:bob",
 	        "relation": "reader",
 	        "object": "document:2021-budget"
 	      },
@@ -196,8 +210,8 @@ type Auth0FgaApi interface {
 	  ]
 	}
 	```
-	This means that `bob@auth0.com` has a `reader` relationship with 1 document `document:2021-budget`.
-	### Query for all users with particular relationships for a particular document
+	This means that `user:bob` has a `reader` relationship with 1 document `document:2021-budget`.
+	### Query for all stored relationship tuples that have a particular relation and object
 	To query for all users that have `reader` relationship with `document:2021-budget`, call read API with body of
 	```json
 	{
@@ -213,7 +227,7 @@ type Auth0FgaApi interface {
 	  "tuples": [
 	    {
 	      "key": {
-	        "user": "bob@auth0.com",
+	        "user": "user:bob",
 	        "relation": "reader",
 	        "object": "document:2021-budget"
 	      },
@@ -222,7 +236,7 @@ type Auth0FgaApi interface {
 	  ]
 	}
 	```
-	This means that `document:2021-budget` has 1 `reader` (`bob@auth0.com`).  Note that the API will not return writers such as `anne@auth0.com` even when all writers are readers.  This is because only direct relationship are returned for the READ API.
+	This means that `document:2021-budget` has 1 `reader` (`user:bob`).  Note that the API will not return writers such as `user:anne` even when all writers are readers.  This is because only direct relationship are returned for the READ API.
 	### Query for all users with all relationships for a particular document
 	To query for all users that have any relationship with `document:2021-budget`, call read API with body of
 	```json
@@ -238,7 +252,7 @@ type Auth0FgaApi interface {
 	  "tuples": [
 	    {
 	      "key": {
-	        "user": "anne@auth0.com",
+	        "user": "user:anne",
 	        "relation": "writer",
 	        "object": "document:2021-budget"
 	      },
@@ -246,7 +260,7 @@ type Auth0FgaApi interface {
 	    },
 	    {
 	      "key": {
-	        "user": "bob@auth0.com",
+	        "user": "user:bob",
 	        "relation": "reader",
 	        "object": "document:2021-budget"
 	      },
@@ -255,7 +269,7 @@ type Auth0FgaApi interface {
 	  ]
 	}
 	```
-	This means that `document:2021-budget` has 1 `reader` (`bob@auth0.com`) and 1 `writer` (`anne@auth0.com`).
+	This means that `document:2021-budget` has 1 `reader` (`user:bob`) and 1 `writer` (`user:anne`).
 
 		 * @param ctx _context.Context - for authentication, logging, cancellation, deadlines, tracing, etc. Passed from http.Request or context.Background().
 		 * @return ApiReadRequest
@@ -270,7 +284,7 @@ type Auth0FgaApi interface {
 
 	/*
 	 * ReadAssertions Read assertions for an authorization model ID
-	 * The GET assertions API will return, for a given authorization model id, all the assertions stored for it. An assertion is an object that contains a tuple key, and the expectation of whether a call to the Check API of that tuple key will return true or false.
+	 * The ReadAssertions API will return, for a given authorization model id, all the assertions stored for it. An assertion is an object that contains a tuple key, and the expectation of whether a call to the Check API of that tuple key will return true or false.
 	 * @param ctx _context.Context - for authentication, logging, cancellation, deadlines, tracing, etc. Passed from http.Request or context.Background().
 	 * @param authorizationModelId
 	 * @return ApiReadAssertionsRequest
@@ -285,17 +299,19 @@ type Auth0FgaApi interface {
 
 	/*
 		 * ReadAuthorizationModel Return a particular version of an authorization model
-		 * The GET authorization-models by ID API will return a particular version of authorization model that had been configured for a certain store.
-	Path parameter `store_id` and `id` are required.
+		 * The ReadAuthorizationModel API returns an authorization model by its identifier.
 	The response will return the authorization model for the particular version.
 
 	## Example
-	To retrieve the authorization model with ID `1yunpF9DkzXMzm0dHrsCuWsooEV` for the store, call the GET authorization-models by ID API with `1yunpF9DkzXMzm0dHrsCuWsooEV` as the `id` path parameter.  The API will return:
+	To retrieve the authorization model with ID `01G5JAVJ41T49E9TT3SKVS7X1J` for the store, call the GET authorization-models by ID API with `01G5JAVJ41T49E9TT3SKVS7X1J` as the `id` path parameter.  The API will return:
 	```json
 	{
 	  "authorization_model":{
-	    "id":"1yunpF9DkzXMzm0dHrsCuWsooEV",
+	    "id":"01G5JAVJ41T49E9TT3SKVS7X1J",
 	    "type_definitions":[
+	      {
+	        "type":"user"
+	      },
 	      {
 	        "type":"document",
 	        "relations":{
@@ -323,7 +339,7 @@ type Auth0FgaApi interface {
 	  }
 	}
 	```
-	In the above example, there is only 1 type (`document`) with 2 relations (`writer` and `reader`).
+	In the above example, there are 2 types (`user` and `document`). The `document` type has 2 relations (`writer` and `reader`).
 		 * @param ctx _context.Context - for authentication, logging, cancellation, deadlines, tracing, etc. Passed from http.Request or context.Background().
 		 * @param id
 		 * @return ApiReadAuthorizationModelRequest
@@ -338,12 +354,41 @@ type Auth0FgaApi interface {
 
 	/*
 		 * ReadAuthorizationModels Return all the authorization model IDs for a particular store
-		 * The GET authorization-models API will return all the IDs of the authorization models for a certain store.
-	Path parameter `store_id` is required.
-	Auth0 FGA's response will contain an array of all authorization model IDs, sorted in descending order of creation.
+		 * The ReadAuthorizationModels API will return all the authorization models for a certain store.
+	Auth0 FGA's response will contain an array of all authorization models, sorted in descending order of creation.
 
-	## [Limits](https://docs.fga.dev/intro/dashboard#limitations)
-	- Each response can contain up to **50** authorization model IDs.
+	## Example
+	Assume that a store's authorization model has been configured twice. To get all the authorization models that have been created in this store, call GET authorization-models. The API will return a response that looks like:
+	```json
+	{
+	  "authorization_models": [
+	    {
+	      "id": "01G50QVV17PECNVAHX1GG4Y5NC",
+	      "type_definitions": [...]
+	    },
+	    {
+	      "id": "01G4ZW8F4A07AKQ8RHSVG9RW04",
+	      "type_definitions": [...]
+	    },
+	  ]
+	}
+	```
+	If there are more authorization models available, the response will contain an extra field `continuation_token`:
+	```json
+	{
+	  "authorization_models": [
+	    {
+	      "id": "01G50QVV17PECNVAHX1GG4Y5NC",
+	      "type_definitions": [...]
+	    },
+	    {
+	      "id": "01G4ZW8F4A07AKQ8RHSVG9RW04",
+	      "type_definitions": [...]
+	    },
+	  ],
+	  "continuation_token": "eyJwayI6IkxBVEVTVF9OU0NPTkZJR19hdXRoMHN0b3JlIiwic2siOiIxem1qbXF3MWZLZExTcUoyN01MdTdqTjh0cWgifQ=="
+	}
+	```
 
 		 * @param ctx _context.Context - for authentication, logging, cancellation, deadlines, tracing, etc. Passed from http.Request or context.Background().
 		 * @return ApiReadAuthorizationModelsRequest
@@ -358,9 +403,9 @@ type Auth0FgaApi interface {
 
 	/*
 		 * ReadChanges Return a list of all the tuple changes
-		 * The GET changes API will return a paginated list of tuple changes (additions and deletions) that occurred in a given store, sorted by ascending time. The response will include a continuation token that is used to get the next set of changes. If there are no changes after the provided continuation token, the same token will be returned in order for it to be used when new changes are recorded. If the store never had any tuples added or removed, this token will be empty.
+		 * The ReadChanges API will return a paginated list of tuple changes (additions and deletions) that occurred in a given store, sorted by ascending time. The response will include a continuation token that is used to get the next set of changes. If there are no changes after the provided continuation token, the same token will be returned in order for it to be used when new changes are recorded. If the store never had any tuples added or removed, this token will be empty.
 	You can use the `type` parameter to only get the list of tuple changes that affect objects of that type.
-	Each store has a limit of **5** requests per second (RPS).
+
 		 * @param ctx _context.Context - for authentication, logging, cancellation, deadlines, tracing, etc. Passed from http.Request or context.Background().
 		 * @return ApiReadChangesRequest
 	*/
@@ -374,20 +419,18 @@ type Auth0FgaApi interface {
 
 	/*
 		 * Write Add or delete tuples from the store
-		 * The POST write API will update the tuples for a certain store.  Tuples and type definitions allow Auth0 FGA to determine whether a relationship exists between an object and an user.
-	Path parameter `store_id` is required.  In the body, `writes` adds new tuples while `deletes` removes existing tuples.
-	## [Limits](https://docs.fga.dev/intro/dashboard#limitations)
-	- Each write API call allows at most **10** tuples.
-	- Each store has a limit of **50000** tuples.
+		 * The Write API will update the tuples for a certain store. Tuples and type definitions allow OpenFGA to determine whether a relationship exists between an object and an user.
+	In the body, `writes` adds new tuples while `deletes` removes existing tuples. The API is not idempotent: if, later on, you try to add the same tuple, or if you try to delete a non-existing tuple, it will throw an error.
+	An `authorization_model_id` may be specified in the body. If it is, it will be used to assert that each written tuple (not deleted) is valid for the model specified. If it is not specified, the latest authorization model ID will be used.
 	## Example
 	### Adding relationships
-	To add `anne@auth0.com` as a `writer` for `document:2021-budget`, call write API with the following
+	To add `user:anne` as a `writer` for `document:2021-budget`, call write API with the following
 	```json
 	{
 	  "writes": {
 	    "tuple_keys": [
 	      {
-	        "user": "anne@auth0.com",
+	        "user": "user:anne",
 	        "relation": "writer",
 	        "object": "document:2021-budget"
 	      }
@@ -396,13 +439,13 @@ type Auth0FgaApi interface {
 	}
 	```
 	### Removing relationships
-	To remove `bob@auth0.com` as a `reader` for `document:2021-budget`, call write API with the following
+	To remove `user:bob` as a `reader` for `document:2021-budget`, call write API with the following
 	```json
 	{
 	  "deletes": {
 	    "tuple_keys": [
 	      {
-	        "user": "bob@auth0.com",
+	        "user": "user:bob",
 	        "relation": "reader",
 	        "object": "document:2021-budget"
 	      }
@@ -424,7 +467,7 @@ type Auth0FgaApi interface {
 
 	/*
 	 * WriteAssertions Upsert assertions for an authorization model ID
-	 * The POST assertions API will add new assertions for an authorization model id, or overwrite the existing ones. An assertion is an object that contains a tuple key, and the expectation of whether a call to the Check API of that tuple key will return true or false.
+	 * The WriteAssertions API will upsert new assertions for an authorization model id, or overwrite the existing ones. An assertion is an object that contains a tuple key, and the expectation of whether a call to the Check API of that tuple key will return true or false.
 	 * @param ctx _context.Context - for authentication, logging, cancellation, deadlines, tracing, etc. Passed from http.Request or context.Background().
 	 * @param authorizationModelId
 	 * @return ApiWriteAssertionsRequest
@@ -438,17 +481,18 @@ type Auth0FgaApi interface {
 
 	/*
 		 * WriteAuthorizationModel Create a new authorization model
-		 * The POST authorization-model API will update the authorization model for a certain store.
-	Path parameter `store_id` and `type_definitions` array in the body are required.  Each item in the `type_definitions` array is a type definition as specified in the field `type_definition`.
+		 * The WriteAuthorizationModel API will add a new authorization model to a store.
+	Each item in the `type_definitions` array is a type definition as specified in the field `type_definition`.
 	The response will return the authorization model's ID in the `id` field.
 
-	## [Limits](https://docs.fga.dev/intro/dashboard#limitations)
-	- There can be at most **24** items in the type_definitions array.
 	## Example
-	To update the authorization model with a single `document` authorization model, call POST authorization-models API with the body:
+	To add an authorization model with `user` and `document` type definitions, call POST authorization-models API with the body:
 	```json
 	{
 	  "type_definitions":[
+	    {
+	      "type":"user"
+	    },
 	    {
 	      "type":"document",
 	      "relations":{
@@ -456,9 +500,7 @@ type Auth0FgaApi interface {
 	          "union":{
 	            "child":[
 	              {
-	                "this":{
-
-	                }
+	                "this":{}
 	              },
 	              {
 	                "computedUserset":{
@@ -470,9 +512,7 @@ type Auth0FgaApi interface {
 	          }
 	        },
 	        "writer":{
-	          "this":{
-
-	          }
+	          "this":{}
 	        }
 	      }
 	    }
@@ -481,7 +521,7 @@ type Auth0FgaApi interface {
 	```
 	Auth0 FGA's response will include the version id for this authorization model, which will look like
 	```
-	{"authorization_model_id": "1yunpF9DkzXMzm0dHrsCuWsooEV"}
+	{"authorization_model_id": "01G50QVV17PECNVAHX1GG4Y5NC"}
 	```
 
 		 * @param ctx _context.Context - for authentication, logging, cancellation, deadlines, tracing, etc. Passed from http.Request or context.Background().
@@ -536,37 +576,36 @@ func (r ApiCheckRequest) Execute() (CheckResponse, *_nethttp.Response, error) {
 
 /*
   - Check Check whether a user is authorized to access an object
-  - The check API will return whether the user has a certain relationship with an object in a certain store.
+  - The Check API queries to check if the user has a certain relationship with an object in a certain store.
 
-Path parameter `store_id` as well as the body parameter `tuple_key` with specified `object`, `relation` and `user` subfields are all required.
-Optionally, a `contextual_tuples` object may also be included in the body of the request. This object contains one field `tuple_keys`, which is an array of tuple keys.
+A `contextual_tuples` object may also be included in the body of the request. This object contains one field `tuple_keys`, which is an array of tuple keys.
+You may also provide an `authorization_model_id` in the body. This will be used to assert that the input `tuple_key` is valid for the model specified. If not specified, the assertion will be made against the latest authorization model ID.
 The response will return whether the relationship exists in the field `allowed`.
 
-## [Limits](https://docs.fga.dev/intro/dashboard#limitations)
 ## Example
-In order to check if user `anne@auth0.com` has a `can_read` relationship with object `document:2021-budget` given the following contextual tuple
+In order to check if user `user:anne` of type `user` has a `reader` relationship with object `document:2021-budget` given the following contextual tuple
 ```json
 
 	{
-	  "user": "anne@auth0.com",
+	  "user": "user:anne",
 	  "relation": "member",
 	  "object": "time_slot:office_hours"
 	}
 
 ```
-a check API call should be fired with the following body:
+the Check API can be used with the following request body:
 ```json
 
 	{
 	  "tuple_key": {
-	    "user": "anne@auth0.com",
-	    "relation": "can_read",
+	    "user": "user:anne",
+	    "relation": "reader",
 	    "object": "document:2021-budget"
 	  },
 	  "contextual_tuples": {
 	    "tuple_keys": [
 	      {
-	        "user": "anne@auth0.com",
+	        "user": "user:anne",
 	        "relation": "member",
 	        "object": "time_slot:office_hours"
 	      }
@@ -859,24 +898,14 @@ func (r ApiExpandRequest) Execute() (ExpandResponse, *_nethttp.Response, error) 
 
 /*
   - Expand Expand all relationships in userset tree format, and following userset rewrite rules.  Useful to reason about and debug a certain relationship
-  - The expand API will return all users (including user and userset) that have certain relationship with an object in a certain store.
+  - The Expand API will return all users and usersets that have certain relationship with an object in a certain store.
 
-This is different from the `/stores/{store_id}/read` API in that both users and computed references are returned.
-Path parameter `store_id` as well as body parameter `object`, `relation` are all required.
-The response will return a userset tree whose leaves are the user id and usersets.  Union, intersection and difference operator are located in the intermediate nodes.
+This is different from the `/stores/{store_id}/read` API in that both users and computed usersets are returned.
+Body parameters `tuple_key.object` and `tuple_key.relation` are all required.
+The response will return a tree whose leaves are the specific users and usersets. Union, intersection and difference operator are located in the intermediate nodes.
 
-## [Limits](https://docs.fga.dev/intro/dashboard#limitations)
 ## Example
-Assume the following type definition for document:
-```yaml
-
-	type document
-	  relations
-	    define reader as self or writer
-	    define writer as self
-
-```
-In order to expand all users that have `reader` relationship with object `document:2021-budget`, an expand API call should be fired with the following body
+To expand all users that have the `reader` relationship with object `document:2021-budget`, use the Expand API with the following request body
 ```json
 
 	{
@@ -887,7 +916,7 @@ In order to expand all users that have `reader` relationship with object `docume
 	}
 
 ```
-Auth0 FGA's response will be a userset tree of the users and computed usersets that have read access to the document.
+Auth0 FGA's response will be a userset tree of the users and usersets that have read access to the document.
 ```json
 
 	{
@@ -901,7 +930,7 @@ Auth0 FGA's response will be a userset tree of the users and computed usersets t
 	            "leaf":{
 	              "users":{
 	                "users":[
-	                  "bob@auth0.com"
+	                  "user:bob"
 	                ]
 	              }
 	            }
@@ -1204,11 +1233,37 @@ func (r ApiListObjectsRequest) Execute() (ListObjectsResponse, *_nethttp.Respons
 }
 
 /*
- * ListObjects [EXPERIMENTAL] Returns a list of all of the object IDs of the provided type that the given user has a specific relation with
- * Please see https://openfga.dev/docs.
- * @param ctx _context.Context - for authentication, logging, cancellation, deadlines, tracing, etc. Passed from http.Request or context.Background().
- * @return ApiListObjectsRequest
- */
+  - ListObjects [EXPERIMENTAL] Get all objects of the given type that the user has a relation with
+  - The ListObjects API returns a list of all the objects of the given type that the user has a relation with. To achieve this, both the store tuples and the authorization model are used.
+
+An `authorization_model_id` may be specified in the body. If it is, it will be used to decide the underlying implementation used. If it is not specified, the latest authorization model ID will be used.
+You may also specify `contextual_tuples` that will be treated as regular tuples.
+The response will contain the related objects in an array in the "objects" field of the response and they will be strings in the object format `<type>:<id>` (e.g. "document:roadmap")
+## Example
+In order to list the objects of type document that user `user:anne` has a `reader` relationship with, while passing the Anne is an editor of the marketing folder in the contextual tuples, You can issue a ListObjects API request that includes the contextual tuples:
+```json
+
+	{
+	  "authorization_model_id": "01G5JAVJ41T49E9TT3SKVS7X1J",
+	  "user": "user:anne",
+	  "relation": "reader",
+	  "type": "document"
+	  "contextual_tuples": {
+	    "tuple_keys": [
+	      {
+	        "user": "user:anne",
+	        "relation": "editor",
+	        "object": "folder:marketing"
+	      }
+	    ]
+	  }
+	}
+
+```
+Auth0 FGA's response will be of the format: `{ "objects": ["document:roadmap"] }` and include document Anne is related to as reader
+  - @param ctx _context.Context - for authentication, logging, cancellation, deadlines, tracing, etc. Passed from http.Request or context.Background().
+  - @return ApiListObjectsRequest
+*/
 func (a *Auth0FgaApiService) ListObjects(ctx _context.Context) ApiListObjectsRequest {
 	return ApiListObjectsRequest{
 		ApiService: a,
@@ -1489,20 +1544,20 @@ func (r ApiReadRequest) Execute() (ReadResponse, *_nethttp.Response, error) {
 
 /*
   - Read Get tuples from the store that matches a query, without following userset rewrite rules
-  - The POST read API will return the tuples for a certain store that matches a query filter specified in the body. Tuples and type definitions allow Auth0 FGA to determine whether a relationship exists between an object and an user.
+  - The Read API will return the tuples for a certain store that match a query filter specified in the body of the request. It is different from the `/stores/{store_id}/expand` API in that it only returns relationship tuples that are stored in the system and satisfy the query.
 
-It is different from the `/stores/{store_id}/expand` API in that only read returns relationship tuples that are stored in the system and satisfy the query.
-It does not expand or traverse the graph by taking the authorization model into account.Path parameter `store_id` is required.  In the body:
-1. Object is mandatory. An object can be a full object (e.g., `type:object_id`) or type only (e.g., `type:`).
-2. User is mandatory in the case the object is type only.
+In the body:
+1. `tuple_key.object` is mandatory. It can be a full object (e.g., `type:object_id`) or type only (e.g., `type:`).
+2. `tuple_key.user` is mandatory in the case the `tuple_key.object` is a type only.
+3. `authorization_model_id` is optional. If specified, it will be used to assert that the input `tuple_key` is valid for the model specified. If not specified, the latest authorization model ID will be used.
 ## Examples
 ### Query for all objects in a type definition
-To query for all objects that `bob@auth0.com` has `reader` relationship in the document type definition, call read API with body of
+To query for all objects that `user:bob` has `reader` relationship in the document type definition, call read API with body of
 ```json
 
 	{
 	 "tuple_key": {
-	     "user": "bob@auth0.com",
+	     "user": "user:bob",
 	     "relation": "reader",
 	     "object": "document:"
 	  }
@@ -1516,7 +1571,7 @@ The API will return tuples and an optional continuation token, something like
 	  "tuples": [
 	    {
 	      "key": {
-	        "user": "bob@auth0.com",
+	        "user": "user:bob",
 	        "relation": "reader",
 	        "object": "document:2021-budget"
 	      },
@@ -1526,8 +1581,8 @@ The API will return tuples and an optional continuation token, something like
 	}
 
 ```
-This means that `bob@auth0.com` has a `reader` relationship with 1 document `document:2021-budget`.
-### Query for all users with particular relationships for a particular document
+This means that `user:bob` has a `reader` relationship with 1 document `document:2021-budget`.
+### Query for all stored relationship tuples that have a particular relation and object
 To query for all users that have `reader` relationship with `document:2021-budget`, call read API with body of
 ```json
 
@@ -1546,7 +1601,7 @@ The API will return something like
 	  "tuples": [
 	    {
 	      "key": {
-	        "user": "bob@auth0.com",
+	        "user": "user:bob",
 	        "relation": "reader",
 	        "object": "document:2021-budget"
 	      },
@@ -1556,7 +1611,7 @@ The API will return something like
 	}
 
 ```
-This means that `document:2021-budget` has 1 `reader` (`bob@auth0.com`).  Note that the API will not return writers such as `anne@auth0.com` even when all writers are readers.  This is because only direct relationship are returned for the READ API.
+This means that `document:2021-budget` has 1 `reader` (`user:bob`).  Note that the API will not return writers such as `user:anne` even when all writers are readers.  This is because only direct relationship are returned for the READ API.
 ### Query for all users with all relationships for a particular document
 To query for all users that have any relationship with `document:2021-budget`, call read API with body of
 ```json
@@ -1575,7 +1630,7 @@ The API will return something like
 	  "tuples": [
 	    {
 	      "key": {
-	        "user": "anne@auth0.com",
+	        "user": "user:anne",
 	        "relation": "writer",
 	        "object": "document:2021-budget"
 	      },
@@ -1583,7 +1638,7 @@ The API will return something like
 	    },
 	    {
 	      "key": {
-	        "user": "bob@auth0.com",
+	        "user": "user:bob",
 	        "relation": "reader",
 	        "object": "document:2021-budget"
 	      },
@@ -1593,7 +1648,7 @@ The API will return something like
 	}
 
 ```
-This means that `document:2021-budget` has 1 `reader` (`bob@auth0.com`) and 1 `writer` (`anne@auth0.com`).
+This means that `document:2021-budget` has 1 `reader` (`user:bob`) and 1 `writer` (`user:anne`).
 
   - @param ctx _context.Context - for authentication, logging, cancellation, deadlines, tracing, etc. Passed from http.Request or context.Background().
   - @return ApiReadRequest
@@ -1873,7 +1928,7 @@ func (r ApiReadAssertionsRequest) Execute() (ReadAssertionsResponse, *_nethttp.R
 
 /*
  * ReadAssertions Read assertions for an authorization model ID
- * The GET assertions API will return, for a given authorization model id, all the assertions stored for it. An assertion is an object that contains a tuple key, and the expectation of whether a call to the Check API of that tuple key will return true or false.
+ * The ReadAssertions API will return, for a given authorization model id, all the assertions stored for it. An assertion is an object that contains a tuple key, and the expectation of whether a call to the Check API of that tuple key will return true or false.
  * @param ctx _context.Context - for authentication, logging, cancellation, deadlines, tracing, etc. Passed from http.Request or context.Background().
  * @param authorizationModelId
  * @return ApiReadAssertionsRequest
@@ -2150,19 +2205,21 @@ func (r ApiReadAuthorizationModelRequest) Execute() (ReadAuthorizationModelRespo
 
 /*
   - ReadAuthorizationModel Return a particular version of an authorization model
-  - The GET authorization-models by ID API will return a particular version of authorization model that had been configured for a certain store.
+  - The ReadAuthorizationModel API returns an authorization model by its identifier.
 
-Path parameter `store_id` and `id` are required.
 The response will return the authorization model for the particular version.
 
 ## Example
-To retrieve the authorization model with ID `1yunpF9DkzXMzm0dHrsCuWsooEV` for the store, call the GET authorization-models by ID API with `1yunpF9DkzXMzm0dHrsCuWsooEV` as the `id` path parameter.  The API will return:
+To retrieve the authorization model with ID `01G5JAVJ41T49E9TT3SKVS7X1J` for the store, call the GET authorization-models by ID API with `01G5JAVJ41T49E9TT3SKVS7X1J` as the `id` path parameter.  The API will return:
 ```json
 
 	{
 	  "authorization_model":{
-	    "id":"1yunpF9DkzXMzm0dHrsCuWsooEV",
+	    "id":"01G5JAVJ41T49E9TT3SKVS7X1J",
 	    "type_definitions":[
+	      {
+	        "type":"user"
+	      },
 	      {
 	        "type":"document",
 	        "relations":{
@@ -2191,7 +2248,7 @@ To retrieve the authorization model with ID `1yunpF9DkzXMzm0dHrsCuWsooEV` for th
 	}
 
 ```
-In the above example, there is only 1 type (`document`) with 2 relations (`writer` and `reader`).
+In the above example, there are 2 types (`user` and `document`). The `document` type has 2 relations (`writer` and `reader`).
   - @param ctx _context.Context - for authentication, logging, cancellation, deadlines, tracing, etc. Passed from http.Request or context.Background().
   - @param id
   - @return ApiReadAuthorizationModelRequest
@@ -2478,13 +2535,46 @@ func (r ApiReadAuthorizationModelsRequest) Execute() (ReadAuthorizationModelsRes
 
 /*
   - ReadAuthorizationModels Return all the authorization model IDs for a particular store
-  - The GET authorization-models API will return all the IDs of the authorization models for a certain store.
+  - The ReadAuthorizationModels API will return all the authorization models for a certain store.
 
-Path parameter `store_id` is required.
-Auth0 FGA's response will contain an array of all authorization model IDs, sorted in descending order of creation.
+Auth0 FGA's response will contain an array of all authorization models, sorted in descending order of creation.
 
-## [Limits](https://docs.fga.dev/intro/dashboard#limitations)
-- Each response can contain up to **50** authorization model IDs.
+## Example
+Assume that a store's authorization model has been configured twice. To get all the authorization models that have been created in this store, call GET authorization-models. The API will return a response that looks like:
+```json
+
+	{
+	  "authorization_models": [
+	    {
+	      "id": "01G50QVV17PECNVAHX1GG4Y5NC",
+	      "type_definitions": [...]
+	    },
+	    {
+	      "id": "01G4ZW8F4A07AKQ8RHSVG9RW04",
+	      "type_definitions": [...]
+	    },
+	  ]
+	}
+
+```
+If there are more authorization models available, the response will contain an extra field `continuation_token`:
+```json
+
+	{
+	  "authorization_models": [
+	    {
+	      "id": "01G50QVV17PECNVAHX1GG4Y5NC",
+	      "type_definitions": [...]
+	    },
+	    {
+	      "id": "01G4ZW8F4A07AKQ8RHSVG9RW04",
+	      "type_definitions": [...]
+	    },
+	  ],
+	  "continuation_token": "eyJwayI6IkxBVEVTVF9OU0NPTkZJR19hdXRoMHN0b3JlIiwic2siOiIxem1qbXF3MWZLZExTcUoyN01MdTdqTjh0cWgifQ=="
+	}
+
+```
 
   - @param ctx _context.Context - for authentication, logging, cancellation, deadlines, tracing, etc. Passed from http.Request or context.Background().
   - @return ApiReadAuthorizationModelsRequest
@@ -2780,10 +2870,10 @@ func (r ApiReadChangesRequest) Execute() (ReadChangesResponse, *_nethttp.Respons
 
 /*
   - ReadChanges Return a list of all the tuple changes
-  - The GET changes API will return a paginated list of tuple changes (additions and deletions) that occurred in a given store, sorted by ascending time. The response will include a continuation token that is used to get the next set of changes. If there are no changes after the provided continuation token, the same token will be returned in order for it to be used when new changes are recorded. If the store never had any tuples added or removed, this token will be empty.
+  - The ReadChanges API will return a paginated list of tuple changes (additions and deletions) that occurred in a given store, sorted by ascending time. The response will include a continuation token that is used to get the next set of changes. If there are no changes after the provided continuation token, the same token will be returned in order for it to be used when new changes are recorded. If the store never had any tuples added or removed, this token will be empty.
 
 You can use the `type` parameter to only get the list of tuple changes that affect objects of that type.
-Each store has a limit of **5** requests per second (RPS).
+
   - @param ctx _context.Context - for authentication, logging, cancellation, deadlines, tracing, etc. Passed from http.Request or context.Background().
   - @return ApiReadChangesRequest
 */
@@ -3071,22 +3161,20 @@ func (r ApiWriteRequest) Execute() (map[string]interface{}, *_nethttp.Response, 
 
 /*
   - Write Add or delete tuples from the store
-  - The POST write API will update the tuples for a certain store.  Tuples and type definitions allow Auth0 FGA to determine whether a relationship exists between an object and an user.
+  - The Write API will update the tuples for a certain store. Tuples and type definitions allow OpenFGA to determine whether a relationship exists between an object and an user.
 
-Path parameter `store_id` is required.  In the body, `writes` adds new tuples while `deletes` removes existing tuples.
-## [Limits](https://docs.fga.dev/intro/dashboard#limitations)
-- Each write API call allows at most **10** tuples.
-- Each store has a limit of **50000** tuples.
+In the body, `writes` adds new tuples while `deletes` removes existing tuples. The API is not idempotent: if, later on, you try to add the same tuple, or if you try to delete a non-existing tuple, it will throw an error.
+An `authorization_model_id` may be specified in the body. If it is, it will be used to assert that each written tuple (not deleted) is valid for the model specified. If it is not specified, the latest authorization model ID will be used.
 ## Example
 ### Adding relationships
-To add `anne@auth0.com` as a `writer` for `document:2021-budget`, call write API with the following
+To add `user:anne` as a `writer` for `document:2021-budget`, call write API with the following
 ```json
 
 	{
 	  "writes": {
 	    "tuple_keys": [
 	      {
-	        "user": "anne@auth0.com",
+	        "user": "user:anne",
 	        "relation": "writer",
 	        "object": "document:2021-budget"
 	      }
@@ -3096,14 +3184,14 @@ To add `anne@auth0.com` as a `writer` for `document:2021-budget`, call write API
 
 ```
 ### Removing relationships
-To remove `bob@auth0.com` as a `reader` for `document:2021-budget`, call write API with the following
+To remove `user:bob` as a `reader` for `document:2021-budget`, call write API with the following
 ```json
 
 	{
 	  "deletes": {
 	    "tuple_keys": [
 	      {
-	        "user": "bob@auth0.com",
+	        "user": "user:bob",
 	        "relation": "reader",
 	        "object": "document:2021-budget"
 	      }
@@ -3397,7 +3485,7 @@ func (r ApiWriteAssertionsRequest) Execute() (*_nethttp.Response, error) {
 
 /*
  * WriteAssertions Upsert assertions for an authorization model ID
- * The POST assertions API will add new assertions for an authorization model id, or overwrite the existing ones. An assertion is an object that contains a tuple key, and the expectation of whether a call to the Check API of that tuple key will return true or false.
+ * The WriteAssertions API will upsert new assertions for an authorization model id, or overwrite the existing ones. An assertion is an object that contains a tuple key, and the expectation of whether a call to the Check API of that tuple key will return true or false.
  * @param ctx _context.Context - for authentication, logging, cancellation, deadlines, tracing, etc. Passed from http.Request or context.Background().
  * @param authorizationModelId
  * @return ApiWriteAssertionsRequest
@@ -3672,19 +3760,20 @@ func (r ApiWriteAuthorizationModelRequest) Execute() (WriteAuthorizationModelRes
 
 /*
   - WriteAuthorizationModel Create a new authorization model
-  - The POST authorization-model API will update the authorization model for a certain store.
+  - The WriteAuthorizationModel API will add a new authorization model to a store.
 
-Path parameter `store_id` and `type_definitions` array in the body are required.  Each item in the `type_definitions` array is a type definition as specified in the field `type_definition`.
+Each item in the `type_definitions` array is a type definition as specified in the field `type_definition`.
 The response will return the authorization model's ID in the `id` field.
 
-## [Limits](https://docs.fga.dev/intro/dashboard#limitations)
-- There can be at most **24** items in the type_definitions array.
 ## Example
-To update the authorization model with a single `document` authorization model, call POST authorization-models API with the body:
+To add an authorization model with `user` and `document` type definitions, call POST authorization-models API with the body:
 ```json
 
 	{
 	  "type_definitions":[
+	    {
+	      "type":"user"
+	    },
 	    {
 	      "type":"document",
 	      "relations":{
@@ -3692,9 +3781,7 @@ To update the authorization model with a single `document` authorization model, 
 	          "union":{
 	            "child":[
 	              {
-	                "this":{
-
-	                }
+	                "this":{}
 	              },
 	              {
 	                "computedUserset":{
@@ -3706,9 +3793,7 @@ To update the authorization model with a single `document` authorization model, 
 	          }
 	        },
 	        "writer":{
-	          "this":{
-
-	          }
+	          "this":{}
 	        }
 	      }
 	    }
@@ -3718,7 +3803,7 @@ To update the authorization model with a single `document` authorization model, 
 ```
 Auth0 FGA's response will include the version id for this authorization model, which will look like
 ```
-{"authorization_model_id": "1yunpF9DkzXMzm0dHrsCuWsooEV"}
+{"authorization_model_id": "01G50QVV17PECNVAHX1GG4Y5NC"}
 ```
 
   - @param ctx _context.Context - for authentication, logging, cancellation, deadlines, tracing, etc. Passed from http.Request or context.Background().